css added

+++ src/Component/Favorite/Favorite.js
@@ -3,7 +3,8 @@ import Navbar from "../../Atom/Navbar/Navbar";
 import CustomButton from "../../Atom/CustomButton/CustomButton";
 import { useNavigate } from "react-router-dom";
 import style from "./Favorite.module.css";
-
+import Swal from 'sweetalert2';
+import swalWithBootstrapButtons from 'sweetalert2';
 export default function Favorite() {
   const previousContacts =
     JSON.parse(localStorage.getItem("favoritePackage")) || [];
@@ -11,7 +12,7 @@ export default function Favorite() {
   const [show, setShow] = useState(false);
   const [allContacts, setAllContacts] = useState(previousContacts);
   const navigate = useNavigate();
-
+ 
   useEffect(() => {
     if (previousContacts.length > 0) {
       setShow(true);
@@ -19,10 +20,40 @@ export default function Favorite() {
   }, [previousContacts]);
 
   const handleDelete = (e) => {
-    const contacts = JSON.parse(localStorage.getItem("favoritePackage")) || [];
-    const otherContacts = contacts.filter((contact) => contact.value1 !== e);
-    localStorage.setItem("favoritePackage", JSON.stringify(otherContacts));
-    setAllContacts(otherContacts);
+
+    swalWithBootstrapButtons
+    .fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "No, cancel!",
+      reverseButtons: true,
+    })
+    .then((result) => {
+      if (result.isConfirmed) {
+        const contacts = JSON.parse(localStorage.getItem("favoritePackage")) || [];
+        const otherContacts = contacts.filter((contact) => contact.value1 !== e);
+        localStorage.setItem("favoritePackage", JSON.stringify(otherContacts));
+        setAllContacts(otherContacts);
+        swalWithBootstrapButtons.fire(
+          "Deleted!",
+          "Your file has been deleted.",
+          "success"
+        );
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          "Cancelled",
+          "Don,t worry Your file is safe :)",
+          "error"
+        );
+      }
+    });
+ 
   };
 
 
@@ -42,7 +73,7 @@ export default function Favorite() {
 
 
   return (
-    <div>
+    <div className={style.main}>
       <Navbar />
       <div className={style.title}>
         <h1>Welcome to Favorite NPM packages</h1>
@@ -64,18 +95,23 @@ export default function Favorite() {
               return (
                 <div className={style.cardmain}>
                   <div className={style.card}>
+                    <div className={style.txtmain}>
+
                     <p>{x.value1}</p>
-                    <p>{x.value2}</p>
+                    </div>
+                    
+                 <div className={style.btnmain}>
                     <CustomButton
                       txt="Edit"
                       onClick={() => HandleEdit(i)}
-                      className={style.btn}
+                      className={style.btn1}
                     />
                     <CustomButton
                       txt="Delete"
                       onClick={() => handleDelete(x.value1)}
-                      className={style.btn}
+                      className={style.btn1}
                     />
+                    </div>
                   </div>
                 </div>
               );
